Unfinished height map work
Created a new HeightMap class to encapsulate the handling of the height
map data.
Added an instance of a height map to the framework to allow the
calculation of ground height for the avatar.
Added a SetOrientation to Loki.
Refactored the LoadTexture helper function and added a new
CreateTextureFromImageData helper.
Updated Volstagg to use the new load texture helper.